Move DSN builder to its own module
